+++ Emulator/CPU/Moira/Moira.h
@@ -234,7 +234,7 @@ protected:
     virtual void signalTraceException() { };
     virtual void signalTrapException() { };
     virtual void signalPrivilegeViolation() { };
-    virtual void signalInterrupt(u8 level) { };
+    virtual void signalInterrupt(int level) { };
     virtual void signalJumpToVector(int nr, u32 addr) { };
 
     // Exception delegates
@@ -353,7 +353,7 @@ private:
     void pollIrq() { reg.ipl = ipl; }
     
     // Selects the IRQ vector to branch to
-    int getIrqVector(int level);
+    u16 getIrqVector(u8 level);
     
 #include "MoiraInit.h"
 #include "MoiraALU.h"
