+++ Emulator/Foundation/Beam.h
@@ -32,7 +32,7 @@ struct Beam
     }
 
     Beam(i16 v, i16 h) : v(v), h(h) { }
-    Beam(u32 cycle = 0) : Beam(cycle / HPOS_CNT, cycle % HPOS_CNT) { }
+    Beam(u32 cycle = 0) : Beam((i16)(cycle / HPOS_CNT), (i16)(cycle % HPOS_CNT)) { }
 
     Beam& operator=(const Beam& beam)
     {
